+++ KeePassium/premium/PricingPlans.swift
@@ -62,7 +62,7 @@ struct PricingPlanCondition {
         case .currentPremiumFeatures:
             return LString.planConditionCurrentPremiumFeatures
         case .perpetualFallback:
-            return LString.planConditionPerpetualFallback
+            return LString.planConditionSubscriptionAsPurchase
         case .familySharing:
             return LString.planConditionFamilySharing
         }
@@ -268,11 +268,11 @@ class PricingPlanPremiumMonthly: RealPricingPlan {
         self.callToAction = LString.premiumCallToActionUpgradeNow
         self.ctaSubtitle = nil
         self.conditions = [
-            PricingPlanCondition(kind: .updatesAndFixes, isIncluded: true, moreInfo: .none),
-            PricingPlanCondition(kind: .emailSupport, isIncluded: true, moreInfo: .none),
             PricingPlanCondition(kind: .allPremiumFeatures, isIncluded: true, moreInfo: .none),
-            PricingPlanCondition(kind: .familySharing, isIncluded: true, moreInfo: .familySharing),
+            PricingPlanCondition(kind: .updatesAndFixes, isIncluded: true, moreInfo: .none),
             PricingPlanCondition(kind: .perpetualFallback, isIncluded: true, moreInfo: .perpetualFallback),
+            PricingPlanCondition(kind: .emailSupport, isIncluded: true, moreInfo: .none),
+            PricingPlanCondition(kind: .familySharing, isIncluded: true, moreInfo: .familySharing),
         ]
         self.benefits = [
             PricingPlanBenefit.multipleDatabases,
@@ -297,11 +297,11 @@ class PricingPlanPremiumYearly: RealPricingPlan {
         self.callToAction = LString.premiumCallToActionUpgradeNow
         self.ctaSubtitle = nil
         self.conditions = [
-            PricingPlanCondition(kind: .updatesAndFixes, isIncluded: true, moreInfo: .none),
-            PricingPlanCondition(kind: .emailSupport, isIncluded: true, moreInfo: .none),
             PricingPlanCondition(kind: .allPremiumFeatures, isIncluded: true, moreInfo: .none),
-            PricingPlanCondition(kind: .familySharing, isIncluded: true, moreInfo: .familySharing),
+            PricingPlanCondition(kind: .updatesAndFixes, isIncluded: true, moreInfo: .none),
             PricingPlanCondition(kind: .perpetualFallback, isIncluded: true, moreInfo: .perpetualFallback),
+            PricingPlanCondition(kind: .emailSupport, isIncluded: true, moreInfo: .none),
+            PricingPlanCondition(kind: .familySharing, isIncluded: true, moreInfo: .familySharing),
         ]
         self.benefits = [
             PricingPlanBenefit.multipleDatabases,
@@ -324,10 +324,10 @@ class PricingPlanPremiumForever: RealPricingPlan {
         self.callToAction = LString.premiumCallToActionUpgradeNow
         self.ctaSubtitle = nil
         self.conditions = [
-            PricingPlanCondition(kind: .updatesAndFixes, isIncluded: true, moreInfo: .none),
-            PricingPlanCondition(kind: .emailSupport, isIncluded: true, moreInfo: .none),
             PricingPlanCondition(kind: .allPremiumFeatures, isIncluded: true, moreInfo: .none),
             PricingPlanCondition(kind: .upcomingPremiumFeatures, isIncluded: true, moreInfo: .none),
+            PricingPlanCondition(kind: .updatesAndFixes, isIncluded: true, moreInfo: .none),
+            PricingPlanCondition(kind: .emailSupport, isIncluded: true, moreInfo: .none),
             PricingPlanCondition(kind: .familySharing, isIncluded: false, moreInfo: .familySharing),
         ]
         self.benefits = [