+++ KeePassium/settings/SettingsVC.swift
@@ -310,10 +310,7 @@ class SettingsVC: UITableViewController, Refreshable {
     }
     
     func didPressManageSubscription() {
-        guard let application = AppGroup.applicationShared,
-            let url = URL(string: "itms-apps://apps.apple.com/account/subscriptions")
-            else { assertionFailure(); return }
-        application.open(url, options: [:])
+        AppStoreHelper.openSubscriptionManagement()
     }