+++ KeePassiumLib/KeePassiumLib/util/DatabaseSettings.swift
@@ -21,6 +21,7 @@ public class DatabaseSettings: Eraseable, Codable {
     public var accessMode: AccessMode
     
     public var isRememberMasterKey: Bool?
+    public var isRememberFinalKey: Bool?
     public private(set) var masterKey: CompositeKey?
     public var hasMasterKey: Bool { return masterKey != nil }
     
@@ -34,6 +35,7 @@ public class DatabaseSettings: Eraseable, Codable {
         case databaseRef
         case accessMode
         case isRememberMasterKey
+        case isRememberFinalKey
         case masterKey
         case isRememberKeyFile
         case associatedKeyFile
@@ -54,6 +56,7 @@ public class DatabaseSettings: Eraseable, Codable {
         self.accessMode = AccessMode.default
         
         isRememberMasterKey = nil
+        isRememberFinalKey = nil
         clearMasterKey()
         
         isRememberKeyFile = nil
@@ -75,6 +78,10 @@ public class DatabaseSettings: Eraseable, Codable {
 
     public func setMasterKey(_ key: CompositeKey) {
         masterKey = key.clone()
+        let isKeepFinalKey = self.isRememberFinalKey ?? Settings.current.isRememberDatabaseFinalKey
+        if !isKeepFinalKey {
+            masterKey?.eraseFinalKeys()
+        }
     }
     
     public func maybeSetMasterKey(_ key: CompositeKey) {
@@ -88,6 +95,10 @@ public class DatabaseSettings: Eraseable, Codable {
         masterKey = nil
     }
 
+    public func clearFinalKey() {
+        masterKey?.eraseFinalKeys()
+    }
+    
     public func setAssociatedKeyFile(_ urlRef: URLReference?) {
         associatedKeyFile = urlRef
     }